python/requirements.txt
oracledb
numpy>=1.20.0
python-dotenv>=0.19.0